# Docs/openapi-alternativerest.yaml
openapi: 3.0.1
info:
  title: Simple i18next service
  description: 'Provides service to add/update i18next resources'
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
tags:
- name: util
  description: privdes basic util functionality of system
paths:
  /i18next/{sysid}/{appid}/{compid}/:
    get:
      tags:
      - util
      summary: gets all messages and labels for app component
      operationId: messages
      parameters:
      - name: sysid
        in: path
        description: system id
        required: true
        schema:
          type: string
      - name: appid
        in: path
        description: application id
        required: true
        schema:
          type: string
      - name: compid
        in: path
        description: component id
        required: true
        schema:
          type: string
      - name: lng
        in: query
        description: language
        required: false
        schema:
          type: string
      - name: nmsp
        in: query
        description: namespace
        required: false
        schema:
          type: string
      responses:
        404:
          description: Resource not found
        200:
          description: OK response
          content:
            application/json:
              schema:
                type: object
                properties:
                  sysid:
                    type: string
                  appid:
                    type: string
                  compid:
                    type: string
                  i18next:
                    type: object
  /i18next/{sysid}/{appid}/{compid}:
    post:
      tags:
      - util
      summary: Adds new application i18next json
      operationId: addMessages
      parameters:
      - name: sysid
        in: path
        description: system id
        required: true
        schema:
          type: string
      - name: appid
        in: path
        description: application id
        required: true
        schema:
          type: string
      - name: compid
        in: path
        description: component id
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/i18next'
      responses:
        201:
          description: Inserted
    put:
      tags:
      - util
      summary: Updates a application i18next json
      operationId: updateMessages
      parameters:
      - name: sysid
        in: path
        description: system id
        required: true
        schema:
          type: string
      - name: appid
        in: path
        description: application id
        required: true
        schema:
          type: string
      - name: compid
        in: path
        description: component id
        required: true
        schema:
          type: string
      responses:
        200:
          description: Updated
    delete:
      tags:
      - util
      summary: Removes all messages for component
      operationId: removeAllMessages
      parameters:
      - name: sysid
        in: path
        description: system id
        required: true
        schema:
          type: string
      - name: appid
        in: path
        description: application id
        required: true
        schema:
          type: string
      - name: compid
        in: path
        description: component id
        required: true
        schema:
          type: string
      responses:
        200:
          description: deleted
          
components:
  schemas:
    i18next:
      title: Instance of a specific i18next language and namespace for the given system/applicaiton/component
      type: object
      properties:
        resID:
          type: integer
          description: id of resource
        sysid:
          type: string
          description: System ID of the resource
        appid:
          type: string
          description: Application ID of the resource
        compid:
          type: string
          description: Component ID of the resource
        language:
          type: string
          description: language locale of resource
        namespace:
          type: string
          description: namespace of the resource
        resource:
          type: string
          description: stringified JSON of the resource file
        modID:
          type: integer
          description: modual id ODOWeb application of the resource
        creTS:
          type: string
          description: creation timestamp of resource
        revTS:
          type: string
          description: revision timestamp of resource
      required:
        - language
        - namepsapce
        - resource
        - sysid
        - appid
        - compid