# service/src/main/java/com/bptopenware/messages/api/I18nextApi.java
/**
 * NOTE: This class is auto generated by the swagger code generator program (3.0.20).
 * https://github.com/swagger-api/swagger-codegen
 * Do not edit the class manually.
 */
package com.bptopenware.messages.api;


import org.springframework.http.ResponseEntity;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestMethod;
import com.bptopenware.messages.i18nextrepository.I18NextDTO;

@javax.annotation.Generated(value = "io.swagger.codegen.v3.generators.java.SpringCodegen", date = "2020-06-26T05:45:41.965Z[GMT]")
public interface I18nextApi {

	@RequestMapping(value = "/i18next/{sysid}/{appid}/{compid}", method = RequestMethod.POST)
	ResponseEntity<Void> addMessages(@PathVariable("sysid") String sysid, @PathVariable("appid") String appid,
			@PathVariable("compid") String compid);

	@RequestMapping(value = "/i18next/{sysid}/{appid}/{compid}/{lng}/{nmsp}", produces = {
			"application/json" }, method = RequestMethod.GET)
	ResponseEntity<I18NextDTO> messages(@PathVariable("sysid") String sysid, @PathVariable("appid") String appid,
			@PathVariable("compid") String compid, @PathVariable("lng") String lng, @PathVariable("nmsp") String nmsp);

	@RequestMapping(value = "/i18next/{sysid}/{appid}/{compid}", method = RequestMethod.DELETE)
	ResponseEntity<Void> removeAllMessages(@PathVariable("sysid") String sysid, @PathVariable("appid") String appid,
			@PathVariable("compid") String compid);

	@RequestMapping(value = "/i18next/{sysid}/{appid}/{compid}", method = RequestMethod.PUT)
	ResponseEntity<Void> updateMessages(@PathVariable("sysid") String sysid, @PathVariable("appid") String appid,
			@PathVariable("compid") String compid);

}